+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+v 5.4.0
+- enabled first version of well-known/split
 v 5.3.9
-- fixed problem with modal not closign after boosting
+- fixed problem with modal not closing after boosting
+
 
 v 5.3.8
 - enabled detection of HiveTube and Podcast 2 plugins